requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0

# Data Analysis and Mathematical Computing
numpy==1.26.2
pandas==2.1.4
scipy==1.11.4
scikit-learn==1.3.2

# Technical Analysis and Financial Calculations
TA-Lib==0.4.28
pandas-ta==0.3.14b0
yfinance==0.2.28
stockstats==0.6.2

# Plotting and Visualization
matplotlib==3.8.2
plotly==5.17.0
seaborn==0.13.0
mplfinance==0.12.10b0

# Database Connectors
PyMySQL==1.1.0
SQLAlchemy==2.0.23
redis==5.0.1

# API and HTTP Libraries
httpx==0.25.2
aiohttp==3.9.1
websockets==12.0

# Telegram Bot
python-telegram-bot==20.7

# Caching and Performance
cachetools==5.3.2
joblib==1.3.2

# Date and Time Handling
python-dateutil==2.8.2
pytz==2023.3

# Configuration and Environment
pydantic==2.5.2
pydantic-settings==2.1.0

# Async Support
asyncio-mqtt==0.16.1
aiofiles==23.2.1

# Utilities
click==8.1.7
colorama==0.4.6
tqdm==4.66.1

# Security and Validation
cryptography==41.0.8
bcrypt==4.1.2

# Testing (for development)
pytest==7.4.3
pytest-asyncio==0.21.1
</txt>
